# docs/open-api/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "DiffScope Synthesis Platform API",
    "version": "0.0.0",
    "description": "DSSP HTTP API surface represented by this service. A server implementation can support a subset of the broader DSSP API."
  },
  "servers": [
    {
      "url": "http://localhost:13711",
      "description": "Default local server URL. The actual host and port are configured by the service."
    }
  ],
...
  "paths": {
...
    "/arch": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchitectureList",
        "summary": "List architectures",
        "description": "Returns metadata for every registered architecture.",
        "parameters": [
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ArchitectureList"
          }
        }
      }
    },
    "/arch/{arch_id}": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchitecture",
        "summary": "Get architecture metadata",
        "description": "Returns metadata for one registered architecture.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchID"
          },
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Architecture"
          }
        }
      }
    },
...
    "/synth/duration": {
      "post": {
        "tags": [
          "Synthesis"
        ],
        "operationId": "postSynthDuration",
        "summary": "Generate phoneme timing",
        "description": "Generates phoneme timing for duration notes. Set stream to true to receive newline-delimited JSON events.",
        "requestBody": {
          "$ref": "#/components/requestBodies/DurationRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/DurationResult"
          }
        }
      }
    },
...
    "/synth/audio": {
      "post": {
        "tags": [
          "Synthesis"
        ],
        "operationId": "postSynthAudio",
        "summary": "Generate audio",
        "description": "Generates encoded audio as a data URL. Set stream to true to receive newline-delimited JSON events.",
        "requestBody": {
          "$ref": "#/components/requestBodies/AudioRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AudioResult"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "ArchID": {
        "name": "arch_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Architecture identifier."
      },
...
      "DisplayLanguage": {
        "name": "display_language",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Preferred language for display metadata."
      }
    },
    "requestBodies": {
...
      "DurationRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DurationRequest"
            }
          }
        }
      },
...
      "AudioRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AudioRequest"
            }
          }
        }
      }
    },
    "responses": {
      "ArchitectureList": {
        "description": "Architecture metadata from every registered architecture.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArchitectureMetadataList"
            }
          }
        }
      },
      "Architecture": {
        "description": "Architecture metadata.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArchitectureMetadata"
            }
          }
        }
      },
...
      "DurationResult": {
        "description": "Duration generation result. Non-streaming requests return a complete JSON response. Streaming requests return newline-delimited JSON progress and completion events.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DurationResponse"
            }
          },
          "application/x-ndjson": {
            "itemSchema": {
              "$ref": "#/components/schemas/DurationStreamEvent"
            }
          }
        }
      },
...
      "AudioResult": {
        "description": "Audio generation result. Non-streaming requests return a complete JSON response. Streaming requests return newline-delimited JSON progress and completion events.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AudioResponse"
            }
          },
          "application/x-ndjson": {
            "itemSchema": {
              "$ref": "#/components/schemas/AudioStreamEvent"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ArchitectureMetadataList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ArchitectureMetadata"
        }
      },
      "ArchitectureMetadata": {
        "type": "object",
        "required": [
          "id",
          "name",
          "pronunciation_mode",
          "phoneme_mode",
          "parameters",
          "audio_dependencies"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Architecture identifier."
          },
          "name": {
            "type": "string"
          },
          "pronunciation_mode": {
            "$ref": "#/components/schemas/PronunciationMode"
          },
          "phoneme_mode": {
            "$ref": "#/components/schemas/PhonemeMode"
          },
          "parameters": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ArchitectureParameterMetadata"
            }
          },
          "audio_dependencies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PronunciationMode": {
        "type": "string",
        "enum": [
          "FULL",
          "SKIP"
        ]
      },
      "PhonemeMode": {
        "type": "string",
        "enum": [
          "FULL",
          "TOKEN_ONLY",
          "SKIP"
        ]
      },
      "ArchitectureParameterMetadata": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/DirectArchitectureParameterMetadata"
          },
          {
            "$ref": "#/components/schemas/IndirectArchitectureParameterMetadata"
          }
        ]
      },
      "DirectArchitectureParameterMetadata": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "const": "DIRECT"
          }
        }
      },
      "IndirectArchitectureParameterMetadata": {
        "type": "object",
        "required": [
          "type",
          "depends_on"
        ],
        "properties": {
          "type": {
            "const": "INDIRECT"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "DurationRequest": {
        "type": "object",
        "required": [
          "context",
          "input"
        ],
        "properties": {
          "context": {
            "$ref": "#/components/schemas/MultiSingerContext"
          },
          "input": {
            "$ref": "#/components/schemas/DurationInput"
          },
          "stream": {
            "type": "boolean",
            "description": "When true, the response uses application/x-ndjson and emits progress events."
          }
        }
      },
...
      "AudioRequest": {
        "type": "object",
        "required": [
          "context",
          "input"
        ],
        "properties": {
          "context": {
            "$ref": "#/components/schemas/MultiSingerContext"
          },
          "input": {
            "$ref": "#/components/schemas/AudioInput"
          },
          "stream": {
            "type": "boolean",
            "description": "When true, the response uses application/x-ndjson and emits progress events."
          }
        }
      },
...
      "MultiSingerContext": {
        "type": "object",
        "required": [
          "arch",
          "arch_extra",
          "singers"
        ],
        "properties": {
          "arch": {
            "type": "string",
            "description": "Architecture identifier."
          },
          "arch_extra": {
            "description": "Architecture-specific JSON payload."
          },
          "singers": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/Singer"
            }
          }
        }
      },
      "Singer": {
        "type": "object",
        "required": [
          "id",
          "extra"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Singer identifier understood by the selected architecture."
          },
          "extra": {
            "description": "Architecture-specific singer JSON payload."
          }
        }
      },
...
      "DurationInput": {
        "type": "object",
        "required": [
          "piece_duration",
          "notes",
          "mix",
          "mix_sample_rate"
        ],
        "properties": {
          "piece_duration": {
            "type": "number",
            "minimum": 0
          },
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DurationNote"
            }
          },
          "mix": {
            "$ref": "#/components/schemas/Mix"
          },
          "mix_sample_rate": {
            "type": "number",
            "exclusiveMinimum": 0
          }
        }
      },
      "DurationNote": {
        "type": "object",
        "required": [
          "position",
          "cent",
          "pronunciation",
          "language",
          "phonemes"
        ],
        "properties": {
          "position": {
            "$ref": "#/components/schemas/NotePosition"
          },
          "cent": {
            "type": "integer",
            "minimum": 0,
            "maximum": 12800
          },
          "pronunciation": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "phonemes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DurationInputPhoneme"
            }
          }
        }
      },
      "NotePosition": {
        "type": "object",
        "required": [
          "gap",
          "duration"
        ],
        "properties": {
          "gap": {
            "type": "number",
            "minimum": 0
          },
          "duration": {
            "type": "number",
            "minimum": 0
          }
        }
      },
      "DurationInputPhoneme": {
        "type": "object",
        "required": [
          "token",
          "onset",
          "language"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "onset": {
            "type": "boolean"
          },
          "language": {
            "type": "string"
          }
        }
      },
...
      "ParameterNote": {
        "type": "object",
        "required": [
          "position",
          "cent",
          "pronunciation",
          "language",
          "phonemes"
        ],
        "properties": {
          "position": {
            "$ref": "#/components/schemas/NotePosition"
          },
          "cent": {
            "type": "integer",
            "minimum": 0,
            "maximum": 12800
          },
          "pronunciation": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "phonemes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParameterInputPhoneme"
            }
          }
        }
      },
      "ParameterInputPhoneme": {
        "type": "object",
        "required": [
          "token",
          "onset",
          "language",
          "start"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "onset": {
            "type": "boolean"
          },
          "language": {
            "type": "string"
          },
          "start": {
            "type": "number"
          }
        }
      },
      "AudioInput": {
        "type": "object",
        "required": [
          "piece_duration",
          "notes",
          "mix",
          "mix_sample_rate",
          "parameters"
        ],
        "properties": {
          "piece_duration": {
            "type": "number",
            "minimum": 0
          },
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParameterNote"
            }
          },
          "mix": {
            "$ref": "#/components/schemas/Mix"
          },
          "mix_sample_rate": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "parameters": {
            "$ref": "#/components/schemas/AudioParameterMap"
          }
        }
      },
      "Mix": {
        "type": "array",
        "description": "Per-frame singer mix data. Each value must be in [0, 1]. Each row must contain exactly singers.length - 1 values, and the sum of values in each row must be less than or equal to 1.",
        "items": {
          "type": "array",
          "items": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          }
        }
      },
...
      "AudioParameterMap": {
        "type": "object",
        "description": "Open string-keyed audio parameter map. Supported keys and semantic requirements are defined by the selected architecture.",
        "additionalProperties": {
          "$ref": "#/components/schemas/AudioParameter"
        }
      },
...
      "AudioParameter": {
        "type": "object",
        "required": [
          "sample_rate"
        ],
        "properties": {
          "values": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "sample_rate": {
            "type": "number",
            "exclusiveMinimum": 0
          }
        }
      },
...
      "DurationResponse": {
        "type": "object",
        "required": [
          "state",
          "output"
        ],
        "properties": {
          "state": {
            "const": "COMPLETE"
          },
          "output": {
            "$ref": "#/components/schemas/DurationOutput"
          }
        }
      },
      "DurationOutput": {
        "type": "object",
        "required": [
          "notes"
        ],
        "properties": {
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DurationOutputNote"
            }
          }
        }
      },
      "DurationOutputNote": {
        "type": "object",
        "required": [
          "phonemes"
        ],
        "properties": {
          "phonemes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DurationOutputPhoneme"
            }
          }
        }
      },
      "DurationOutputPhoneme": {
        "type": "object",
        "required": [
          "start"
        ],
        "properties": {
          "start": {
            "type": "number"
          }
        }
      },
...
      "AudioResponse": {
        "type": "object",
        "required": [
          "state",
          "output"
        ],
        "properties": {
          "state": {
            "const": "COMPLETE"
          },
          "output": {
            "$ref": "#/components/schemas/AudioOutput"
          }
        }
      },
      "AudioOutput": {
        "type": "object",
        "required": [
          "audio_url"
        ],
        "properties": {
          "audio_url": {
            "type": "string",
            "description": "Encoded audio data URL."
          }
        }
      },
...
      "DurationStreamEvent": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/StreamStateEvent"
          },
          {
            "$ref": "#/components/schemas/DurationResponse"
          }
        ]
      },
...
      "AudioStreamEvent": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/StreamStateEvent"
          },
          {
            "$ref": "#/components/schemas/AudioResponse"
          }
        ]
      },
      "StreamStateEvent": {
        "type": "object",
        "required": [
          "state"
        ],
        "properties": {
          "state": {
            "enum": [
              "QUEUING",
              "PROCESSING"
            ]
          }
        }
      },
...
    }
  }
}